refactor: Extract the top menu components into separate files
