+++ src/app/(frontend)/articles/[slug]/page.tsx
@@ -8,7 +8,7 @@ import { generateMeta } from '@/utilities/generateMeta'
 import { Metadata } from 'next'
 import { LivePreviewListener } from '@/components/LivePreviewListener'
 import RelatedArticles from '@/components/Article/RelatedArticles'
-import TopMenuContainer from '@/components/Article/TopMenuContainer'
+import TopMenuContainer from '@/components/Article/TopMenu/TopMenuContainer'
 import RightMenuContainer from '@/components/Article/RightMenuContainer'
 import BottomMenu from '@/components/Article/BottomMenu'
 import { ArticleHero } from '@/heros/ArticleHero'
